compute_technical_interactions {icesFO}	R Documentation

Compute a matrix of technical interactions.

Description

Based on landings and catches by species and gear, compute a matrix of technical interaction values and main gear contributions.

Usage

compute_technical_interactions(x, threshByCatch = 5,
  thresholdMainGear = 0.8, thresholdCoCatches = 0.5,
  catchCoverage = 0.95)

Arguments

x
a dataframe in the same format as the STECF data.

threshByCatch
minimum tonnage for a species to be considered a by catch (default 5)

thresholdMainGear
a threshold for including gears contributing to co-catches, to retain only the gear with the greatest contribution set this to 1 (default is 0.8)

thresholdCoCatches
a proportion giving the threshold for when we decide that there is co-catches (default 0.5)

catchCoverage
the propotion of total catch to cover in the tables i.e. 0.95 gives a table convering 95% of the total catch (default 0.95)

Value

A list with four elements, technical interactions for landings (recapLand) and catch (recapCatch) and tables of main gears involved in the interactions based on landings (MainGearsLandings) and catches (MainGearsCatches)

Note

Technical interaction appears between stocks when they are caught by the same gear during a fishing operation. Ideally the technical interaction should then be studied at the scale of the fishing operation to prevent artificially creating technical interaction between stocks that might only be caught at day/night or in different areas/timing of the year. However, often, the finest available information is per stock/gear/area/season. Knowing these limitations, the methodology used here consists in computing the sum of landings per strata of one species given that a second species is also present in the landings of this strata and then dividing this number by the total landings of the first species:

T(i,j)=(sum_s L(s,j)*P(s,j)) / (sum_s L(s,i)) *100

Where T is the value of the technical interaction, i and j are the two species for which the technical interaction is assessed. P is an indicator variable and equals 0 if the total landings of a species for a for a given strata is less than 5 total landings for that strata and 1 otherwise. L is the landings for a given species and strata.

Strata corresponds to the provided disaggregation of the landings.

See Also

plot_technical_interactions for plotting technical interactions.

icesFO-package gives an overview of the package.

Examples

## Not run: 
technical_interacton <- 
  compute_technical_interactions(STECF_landings, catchCoverage = 0.99)

plot_technical_interactions(technical_interacton$recapLand)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]