format_annex_table {icesFO}	R Documentation

Format the SAG status data for its use as an annex table.

Description

It provides an excel file with color names per status, as well as other information that might be required for an annex table

Usage

format_annex_table(df, year)

Arguments

df
a dataframe output from format_sag_status.

year
the active year

Value

A data frame

Note

Can add some helpful information here

See Also

format_sag for formatting raw data from the ICES Stock Assessment database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
sid_format <- format_sid(x, "Celtic Seas")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]