format_catches {icesFO}	R Documentation

Format the data from the time-series of ICES historical, official and preliminary catches.

Description

Format the data from the ICES Stock Assessment Database for the downloaded year and the specific Ecoregion for which you are producing the Fisheries Overviews.

Usage

format_catches(year, ecoregion, historical, official, preliminary = NULL,
  species_list, sid)

Arguments

year
the year required

ecoregion
an identifier of the Ecoregion of interest

historical
a dataframe output from load_historical_catches() required.

official
a dataframe output from load_official_catches() required.

preliminary
a dataframe output from load_preliminary_catches() optional.

species_list
a list of species common names and scientific names from ASFIS

sid
the main table from ICES Stock Information DB (SD)

Value

A data frame..

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

format_sid for formatting raw data from the ICES Stock Information database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
  sid <- load_sid(2019)
  species_list <- load_asfis_species()
  historical <- load_historical_catches()
  official <- load_official_catches()
  preliminary <- load_preliminary_catches(2018)
  catches_format <- 
    format_catches(2019, "Celtic Seas", historical, official, preliminary,
                   species_list, sid)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]