format_sag {icesFO}	R Documentation

Format the data from the ICES Stock Assessment Database.

Description

Format the data from the ICES Stock Assessment Database for the downloaded year and the specific Ecoregion for which you are producing the Fisheries Overviews.

Usage

format_sag(x, y, year, ecoregion)

Arguments

x
a dataframe output from load_sag_summary() required.

y
a dataframe output from load_sag_refpts() required.

year
the year for which data is required.

ecoregion
an identifier of the Ecoregion of interest

Value

A data frame..

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

format_sid for formatting raw data from the ICES Stock Information database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
sag_format <- format_sag(x,y, "Celtic Seas")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]