format_sid {icesFO}	R Documentation

Format the data from the ICES Stock Information Database.

Description

Format the data from the ICES Stock Information Database for the downloaded year and the specific Ecoregion for which you are producing the Fisheries Overviews.

Usage

format_sid(x, ecoregion)

Arguments

x
a dataframe output from load_sid() required.

ecoregion
an identifier of the Ecoregion of interest

Value

A data frame..

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

format_sag for formatting raw data from the ICES Stock Assessment database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
sid_format <- format_sid(x, "Celtic Seas")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]