format_stecf {icesFO}	R Documentation

Format the data from STECF effort and landings

Description

Format the data from STECF effort and landings for the specific Ecoregion for which you are producing the Fisheries Overviews.

Usage

format_stecf_effort(x)

format_stecf_landings(x)

Arguments

x
the name of the dataframe with effort data

Details

These two dataframes have to be downloaded by hand and put in the data folder. The proper Annexes have to decided by the user.

Value

a data frame of stock status relative to reference points and catch, discards, and landings by stock for the most recent assessment.

Note

Some considerable errors have been identified in the STECF data. Finland and Estonia effort data are not reliable, and Germany recorded an erroneous haul in 2013. These values have been removed.

Can add some helpful information here

References

STECF dissemination tool https://stecf.jrc.ec.europa.eu/web/stecf/dd/effort/graphs-annex

See Also

format_sag for formatting raw data from the ICES Stock Assessment database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
stecf_formatted <- format_stecf("Celtic Seas")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]