load_catches {icesFO}	R Documentation

Download catch data from ICES web services.

Description

Download historical and current catches data from the ICES web services also preliminary catches can be downloaded for the current year

Usage

load_historical_catches()

load_official_catches()

load_preliminary_catches(year)

Arguments

year
the year for which data is required.

Value

A data frame..

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

load_sag for loading data from the ICES Stock Assessment database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
catches_hist_raw <- load_historical_catches()
catches_official_raw <- load_official_catches()
catches_prelim_raw <- load_preliminary_catches()

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]