load_ecoregion {icesFO}	R Documentation

Download ecoregion polygons

Description

Returns a simple features object with a polygon for each ecoregion

Usage

load_ecoregion(ecoregion, precision = 3)

Arguments

ecoregion
an ICES ecoregion to download (e.g "Baltic Sea")

precision
the numnber of decimal places required in the coordinates

Value

A simple features collection

See Also

icesFO-package gives an overview of the package.

Examples

## Not run: 
  ecoregion <- load_ecoregion("Baltic Sea")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]