plot_catch_trends {icesFO}	R Documentation

Returns an ordered plot of catch bars colored according to F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub> by fish category and ecoregion

Description

Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for each stock in the Ecoregion, according to the last assessment (relative to the set year)

Usage

plot_catch_trends(x, type = c("COMMON_NAME", "COUNTRY", "GUILD"),
  line_count = 10, plot_type = c("line", "area"),
  preliminary_catches = TRUE, official_catches_year = 2018,
  return_data = FALSE)

Arguments

x
a dataframe output of stockstatus+catch_current.R

type
COMMON_NAME, COUNTRY or GUILD

line_count
number of lines to show

plot_type
either line or area

preliminary_catches
logical flag

official_catches_year
year required

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A plot

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

plot_CLD_bar Stock status relative to reference points.

icesFO-package gives an overview of the package.

Examples

## Not run: 
plot1 <- plot_CLD_bar(CLDstatus, guild="Demersal", return_data = TRUE)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]