plot_discard_trends {icesFO}	R Documentation

Returns a plot of discard rate trends by fish category for an ecoregion.

Description

Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for each stock in the Ecoregion, according to the last assessment (relative to the set year)

Usage

plot_discard_trends(x, year, caption = FALSE, cap_year, cap_month,
  return_data = FALSE)

Arguments

x
a dataframe output of CLD_trends

year
year required

caption
logial flag

cap_year
the year to be shown in the figure caption

cap_month
the month to be shown in the figure caption, the accession date to SAG usually

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A plot

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

plot_discard_current a plot of landings and discards by fish category in the last year for an ecoregion.

icesFO-package gives an overview of the package.

Examples

## Not run: 
plot1 <- plot_discard_trends(CLDtrends, caption = T,cap_year, cap_month, return_data = TRUE)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]