plot_ecoregion_map {icesFO}	R Documentation

Returns an map of the ecoregion and the divisions included in it

Description

Returns an map of the ecoregion and the divisions included in it

Usage

plot_ecoregion_map(ecoregion, ices_areas)

Arguments

ecoregion
a simple features dataframe output of load_ecoregions( ... )

ices_areas
a simple features dataframe output of load_areas( ... )

Value

A ggplot plot

See Also

plot_CLD_bar Stock status relative to reference points.

icesFO-package gives an overview of the package.

Examples

## Not run: 
  ices_areas <- load_areas("Baltic Sea")
  ecoregion <- load_ecoregion("Baltic Sea")
  eco_map <- plot_ecoregion_map(ecoregion, ices_areas)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]