plot_guild_trends {icesFO}	R Documentation

Plotting time series of F/Fmsy and SSB/MSY B trigger by guilds. This plots are used int the Ecosystem Overviews.

Description

Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for each guild in the Ecoregion, according to the last assessment (relative to the set year)

Usage

plot_guild_trends(x, cap_year, cap_month, return_data = FALSE)

Arguments

x
a dataframe output of guild_trends function

cap_year
the year to be shown in the figure caption

cap_month
the month to be shown in the figure caption, the accession date to SAG usually

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A plot

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

format_sag for formatting data from the ICES Stock Assessment database.

icesFO-package gives an overview of the package.

Examples

## Not run: 
plot1 <- plot_guild_trends(trends2018)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]