plot_sar_map {icesFO}	R Documentation

Returns an map of the ecoregion and the divisions included in it

Description

Returns an map of the ecoregion and the divisions included in it

Usage

plot_sar_map(sar, ecoregion, what)

Arguments

sar
a dataframe of swept area ratio data with a WKT columns (see notes)

ecoregion
a dataframe output of load_ecoregion( ... )

what
a flag which SAR value to plot - either "surface" or "subsurface"

Value

A ggplot object

Note

The effort data.frame is conveted to an sf object for plotting and is expeted to have a Well Known Text column containing polygons of c-squares

See Also

plot_ecoregion_map plots ICES areas and ecoregion.

Examples

## Not run: 
ecoregion <- load_ecoregion("Baltic Sea")
sar <- icesVMS::get_sar_map("Baltic Sea")

# convert to sf
sar <- sf::st_as_sf(sar, wkt = "wkt", crs = 4326)

plot1 <- plot_sar_map(sar, ecoregion, what = "surface")

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]