plot_status_prop_pies {icesFO}	R Documentation

Returns an ordered plot of catch bars colored according to F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub> by fish category and ecoregion

Description

Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for each stock in the Ecoregion, according to the last assessment (relative to the set year)

Usage

plot_status_prop_pies(x, cap_month = "November", cap_year = "2018",
  return_data = FALSE)

Arguments

x
a dataframe output of format_sag_status.R

cap_month
the month to be shown in the figure caption, the accession date to SAG usually

cap_year
the year to be shown in the figure caption

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A plot

Note

Can add some helpful information here

References

The ICES stock information Database web sevices: http://sid.ices.dk/services/

See Also

plot_CLD_bar Stock status relative to reference points.

icesFO-package gives an overview of the package.

Examples

## Not run: 
plot1 <- plot_CLD_bar(CLDstatus, guild="Demersal", return_data = TRUE)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]