plot_stecf {icesFO}	R Documentation

STECF Landings over time by country, guild, or species

Description

The stecf_plot function returns an area or line plot of landings (historic and official catch) for an ecoregion by country guild, or species.

Usage

plot_stecf(x, type, variable = NULL, cap_year, cap_month, line_count,
  stecf_report, return_data = FALSE)

Arguments

x
a dataframe resulting from format_stecf_effort function

type
need description

variable
need description

cap_year
the year to be shown in the figure caption

cap_month
the month to be shown in the figure caption, the accession date to SAG usually

line_count
indicates the number of lines to be shown in the graph

stecf_report
need description

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A ggplot2 object

Examples

## Not run: 
stecf_plot("Greater North Sea Ecoregion", 
           metric = "EFFORT", 
           type = "GEAR", 
           return_plot = TRUE, 
           line_count = 4)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]