plot_technical_interactions {icesFO}	R Documentation

Visualise a matrix of technical interactions.

Description

Based on a matrix of technical interactions produce an image plot with a legend.

Usage

plot_technical_interactions(x, col = heat.colors(5, rev = TRUE))

Arguments

x
a square matrix with identical row and column names containing values of tecnical interaction between species

col
a vector of colours to use on the image plot

Value

A list with four elements, technical interactions for landings (recapLand) and catch (recapCatch) and tables of main gears involved in the interactions based on landings (MainGearsLandings) and catches (MainGearsCatches)

See Also

compute_technical_interactions for computing technical interactions.

icesFO-package gives an overview of the package.

Examples

## Not run: 
technical_interacton <- 
  compute_technical_interactions(STECF_landings, catchCoverage = 0.99)

plot_technical_interactions(technical_interacton$recapLand)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]