plot_vms {icesFO}	R Documentation

VMS Effort and landings over time by country or gear

Description

The stecf_plot function returns an area or line plot of landings (historic and official catch) for an ecoregion by country guild, or species.

Usage

plot_vms(x, metric = NULL, type = NULL, cap_year, cap_month,
  line_count, return_data = FALSE)

Arguments

x
a dataframe resulting from icesVMS::get_fo_effort() or icesVMS::get_fo_landings()

metric
either "effort" or "landings"

type
either "country" or "gear_category"

cap_year
the year to be shown in the figure caption

cap_month
the month to be shown in the figure caption, the accession date to SAG usually

line_count
indicates the number of lines to be shown in the graph

return_data
a parameter indicating if the data behind the plot should be returned as a dataframe

Value

A ggplot2 object

Examples

## Not run: 
vms_plot("Greater North Sea Ecoregion", 
           metric = "effort", 
           type = "gear_category",
           line_count = 4)

## End(Not run)

[Package icesFO version 1.0.0.9003 Index]